改进 管理员控制台命令。

+++ app/Console/Commands/Admin/Delete.php
@@ -20,7 +20,7 @@ class Delete extends Command
      *
      * @var string
      */
-    protected $description = '删除一个管理员';
+    protected $description = '删除一个管理员。';
 
     /**
      * Execute the console command.
@@ -30,7 +30,29 @@ class Delete extends Command
     public function handle(): int
     {
         // 获取管理员ID
-        $id = $this->ask('请输入管理员ID');
+        $id = $this->ask('请输入管理员 ID');
+
+        // 搜索
+        $admin = Admin::find($id);
+        if (is_null($admin)) {
+            $this->error('管理员不存在。');
+
+            return CommandAlias::FAILURE;
+        }
+
+        // 输出信息
+        $this->table(['ID', '名称', '邮箱'], [
+            [
+                'id' => $admin->id,
+                'name' => $admin->name,
+                'email' => $admin->email,
+            ],
+        ]);
+
+        // 确认
+        if (! $this->confirm('确认删除管理员吗？')) {
+            return CommandAlias::FAILURE;
+        }
 
         // 删除管理员
         Admin::destroy($id);
